// apps/admin/src/api/generated/schemas/apiKeyCreate.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ApiKeyCreateName } from './apiKeyCreateName';
import type { ApiKeyCreateScopes } from './apiKeyCreateScopes';
import type { ApiKeyCreateRateLimit } from './apiKeyCreateRateLimit';

export interface ApiKeyCreate {
  tenantId: string;
  name?: ApiKeyCreateName;
  scopes?: ApiKeyCreateScopes;
  rateLimit?: ApiKeyCreateRateLimit;
}